apply to children
